+++ changelog.md
@@ -9,6 +9,7 @@ _Date TBD_
 
 **New Features:**
 
+* Add `flatMapIndexed` specializations for flattening immutable arrays
 * Add `flatMap` specializations for flattening immutable arrays
 * Add `requireNoNulls` extension function
 * Add `indexOfLast(predicate)` method